refactor: move version back to root
