+++ main.go
@@ -4,7 +4,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 	"github.com/terraform-providers/terraform-provider-selectel/selectel"
-	"github.com/terraform-providers/terraform-provider-selectel/selectel/version"
+	"github.com/terraform-providers/terraform-provider-selectel/version"
 )
 
 func main() {
